Creación métodos y clases de persistencia

+++ EFT_S9/src/main/java/com/duoc/eft_s9/model/Vehiculo.java
@@ -6,10 +6,10 @@ public abstract class Vehiculo {
     protected String patente;
     protected String marca;
     protected int anio;
-    protected double precioBase;
+    protected int precioBase;
     protected TipoVehiculo tipoVehiculo;
 
-    public Vehiculo(String patente, String marca, int anio, double precioBase, TipoVehiculo tipoVehiculo) {
+    public Vehiculo(String patente, String marca, int anio, int precioBase, TipoVehiculo tipoVehiculo) {
         this.patente = patente;
         this.marca = marca;
         this.anio = anio;
@@ -17,26 +17,26 @@ public abstract class Vehiculo {
         this.tipoVehiculo = tipoVehiculo;
     }
 
-    public abstract double calcularPrecioFinal();
+    public abstract int calcularPrecioFinal();
 
     // Getters
     public String getPatente() { return patente; }
     public String getMarca() { return marca; }
     public int getAnio() { return anio; }
-    public double getPrecioBase() { return precioBase; }
+    public int getPrecioBase() { return precioBase; }
     public TipoVehiculo getTipoVehiculo() { return tipoVehiculo; }
 
     // Setters
     public void setMarca(String marca) { this.marca = marca; }
     public void setAnio(int anio) { this.anio = anio; }
-    public void setPrecioBase(double precioBase) { this.precioBase = precioBase; }
+    public void setPrecioBase(int precioBase) { this.precioBase = precioBase; }
     public void setTipoVehiculo(TipoVehiculo tipoVehiculo) {
         this.tipoVehiculo = tipoVehiculo;
     }
 
     @Override
     public String toString() {
-        return String.format("Patente: %s | Marca: %s | Año: %d | Tipo: %s | Precio Base: %.2f",
+        return String.format("Patente: %s | Marca: %s | Año: %d | Tipo: %s | Precio Base: %d",
                 patente, marca, anio, tipoVehiculo.name(), precioBase);
     }
 }
