+++ EFT_S9/src/main/java/com/duoc/eft_s9/model/VehiculoCarga.java
@@ -3,10 +3,10 @@ package com.duoc.eft_s9.model;
 import com.duoc.eft_s9.enums.TipoVehiculo;
 
 public class VehiculoCarga extends Vehiculo {
-    private double capacidadToneladas;
+    private int capacidadToneladas;
 
-    public VehiculoCarga(String patente, String marca, int anio, double precioBase,
-                         TipoVehiculo tipoVehiculo, double capacidadToneladas) {
+    public VehiculoCarga(String patente, String marca, int anio, int precioBase,
+                         TipoVehiculo tipoVehiculo, int capacidadToneladas) {
         super(patente, marca, anio, precioBase, tipoVehiculo);
         if (tipoVehiculo.getCategoria() != TipoVehiculo.Categoria.CARGA) {
             throw new IllegalArgumentException("TipoVehiculo no válido para VehiculoCarga.");
@@ -15,14 +15,14 @@ public class VehiculoCarga extends Vehiculo {
     }
 
     @Override
-    public double calcularPrecioFinal() {
-        double extra = capacidadToneladas * 25000;
+    public int calcularPrecioFinal() {
+        int extra = capacidadToneladas * 25000;
         return precioBase + extra;
     }
 
     // Getters y Setters
-    public double getCapacidadToneladas() { return capacidadToneladas; }
-    public void setCapacidadToneladas(double capacidadToneladas) {
+    public int getCapacidadToneladas() { return capacidadToneladas; }
+    public void setCapacidadToneladas(int capacidadToneladas) {
         this.capacidadToneladas = capacidadToneladas;
     }
 }
